shooting_range_front/requirements.txt
amqp==5.1.1
async-timeout==4.0.2
bcrypt==3.2.0
billiard==3.6.4.0
blinker==1.4
celery==5.2.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cryptography==36.0.2
Deprecated==1.2.13
Flask==2.1.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.0
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
kombu==5.2.4
MarkupSafe==2.1.1
marshmallow==3.15.0
packaging==21.3
prompt-toolkit==3.0.29
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.8
python-dotenv==0.20.0
python-slugify==6.1.1
pytz==2022.1
redis==4.2.2
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.34
text-unidecode==1.3
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.1.1
wrapt==1.14.0
WTForms==3.0.1
